Factor out openapi types
