+++ packages/types/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./api.js";
+export * from "./openapi.js";
 export * from "./otel.js";
 export * from "./schemas.js";
 export * from "./settings.js";
